Create comments when the build is done
This will allow us to respect the MAX_COMMENTS environment variable
across the whole build and should help alleviate some abuse warnings.

As-is, if an external worker fails for whatever reason, the build will
not be marked as complete and comments will never go out. I *think*
we'll want to get that working by having the workers be retryable and we
can mark the build as complete if it fails too many times. Maybe this is
good enough for now, though?

An alternative approach would be to add this logic to `BuildReport`
itself, when can bail early if the build is not complete.